Simplify zope.event handling
This change greatly simplifies how we handle zope.event events. Before, we
needed to add a new subscriber to `events.zcml`, and a handler method to
`events.py`. Consequently, the `events.py` was a toilet for all event handlers
and could grow to be very unreadable. What we do from now on is add event
handlers into the `content/` folder, into the file of the content item the event
is dealing with. Moreover, we replace the use of `events.zcml` for registering
event subscribers to using `grok.subscribe` method decorator. So no *ZCML*,
everything is done in one file, with python.