Don't use version.txt any more. Use version in setup.py.
Back in the days, Plone add-ons used version.txt to specify their version.
Then eggs came and version was not specified in setup.py. For some time we
merged the two approaches to read version.txt into setup.py. Today, this is not
needed any more as Plone is fully egg-ified and we can use just the setup.py.