Fixes #31 Sane fields for Events
Probably it will be unnecessary once we use the standar Plone events, but JIC.

Note that I removed the IBasic behavior because our description richtext and I
didn't manage to override the default one.

I also needed to recreate my var folder to apply the changes. Probably there's
a fancier way to do it.